Updates to libregex to allow the mged/archer 'red' command to work on 64bit windows. Updated files in 'src/other/libregex/' ...  'regex.h' 'regex2.h' 'regexec.c' and 'engine.c'.
svn:revision:51054
svn:branch:trunk
svn:account:r_weiss

+++ src/other/libregex/engine.c
@@ -1038,7 +1038,7 @@ sopno stopst;
 
 	printf("%s %s-", title, pchar(*start));
 	printf("%s ", pchar(*stop));
-	printf("%ld-%ld\n", (long)startst, (long)stopst);
+	printf("%ld-%ld\n", (ssize_t)startst, (ssize_t)stopst);
 }
 
 #ifndef PCHARDONE
